refactor: 에러 UI 렌더링이 fetch가 호출되는 곳에서 실행되도록 수정
