+++ src/api/movieApi.ts
@@ -1,7 +1,5 @@
 import api from "./api.ts";
 import { OPTION } from "../constants/api.ts";
-import { ERROR, STATUS_MESSAGE } from "../constants/error.ts";
-import ErrorUI from "../components/ErrorUI.ts";
 
 const defaultParams = {
   language: OPTION.language,
@@ -20,11 +18,7 @@ const movieApi = {
       return await api.GETWithAuth(endpoint);
     } catch (error) {
       if (error instanceof Error) {
-        const status = Number(error.message);
-        const message = STATUS_MESSAGE[status] ?? ERROR.DEFAULT;
-        const errorUI = new ErrorUI({ status, message });
-        errorUI.create();
-        errorUI.renderError();
+        throw error;
       }
     }
   },
@@ -41,11 +35,7 @@ const movieApi = {
       return await api.GETWithAuth(endpoint);
     } catch (error) {
       if (error instanceof Error) {
-        const status = Number(error.message);
-        const message = STATUS_MESSAGE[status] ?? ERROR.DEFAULT;
-        const errorUI = new ErrorUI({ status, message });
-        errorUI.create();
-        errorUI.renderError();
+        throw error;
       }
     }
   },
@@ -60,11 +50,7 @@ const movieApi = {
       return await api.GETWithAuth(endpoint);
     } catch (error) {
       if (error instanceof Error) {
-        const status = Number(error.message);
-        const message = STATUS_MESSAGE[status] ?? ERROR.DEFAULT;
-        const errorUI = new ErrorUI({ status, message });
-        errorUI.create();
-        errorUI.renderError();
+        throw error;
       }
     }
   },
